feat: Adição do tratamento de exceção para not found e adição no swagger para o método de delete de entregas canceladas

+++ delivery/src/main/java/com/esand/delivery/service/DeliveryService.java
@@ -134,6 +134,9 @@ public class DeliveryService {
 
     @Transactional
     public void deleteAllCanceled() {
+        if (!deliveryRepository.existsByStatus(Delivery.Status.CANCELED)) {
+            throw new EntityNotFoundException("No deliveries canceled found");
+        }
         deliveryRepository.deleteAllByStatus(Delivery.Status.CANCELED);
     }
